add some functional tests (rcov ~55%)

+++ app/controllers/person_sessions_controller.rb
@@ -10,13 +10,14 @@ class PersonSessionsController < ApplicationController
       flash[:notice] = t('log.in.success').capitalize
     else
       flash[:error] = t('log.in.failed').capitalize
-      redirect_to_stored_location({:action => new})
+      redirect_to_stored_location([forgeos_core, :login])
     end
   end
 
   def destroy
-    current_user_session.destroy
-    flash[:notice] = I18n.t('log.out.success').capitalize
+    if current_user_session and current_user_session.destroy
+      flash[:notice] = I18n.t('log.out.success').capitalize
+    end
     redirect_to(:root)
   end
 end
